requirements.txt
# Core
pyyaml==6.0.1

# Data processing
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.4

# Web scraping / API
requests==2.31.0
pytrends==4.9.2
zstandard==0.22.0

# Analysis
scipy==1.12.0
scikit-learn==1.4.0
ruptures==1.1.9
statsmodels==0.14.1

# Visualization
matplotlib==3.8.3
seaborn==0.13.2

# Pipeline
click==8.1.7
tqdm==4.66.2
tenacity==8.2.3

# Document generation
python-docx==1.1.0

# Optional: GCP ingestion (pip install -r requirements-gcp.txt)
# pyspark==3.5.0
# google-cloud-bigquery==3.25.0
# google-cloud-storage==2.18.0
# google-cloud-dataproc==5.9.0
# db-dtypes==1.2.0
# warcio==1.7.4
